refactor(desktop): usage 哨兵解析归入 services/pr-agent
services/usage.ts 解析 pr-agent 嵌入式 shim 注入的 @@MEEBOX_USAGE@@ 哨兵行（USAGE_SENTINEL）+ 累加
token，是 pr-agent 域专属（非通用 util），消费方均在 pr-agent 域（run-executor + 编排 chat）。移入
services/pr-agent/usage.ts：run-executor 改同目录 ./usage.js，orchestrator 改 ../pr-agent/usage.js。

纯移动、行为不变；desktop typecheck/lint/build 全绿。
